CORE-8910 Remove old code and merge existing code where possible
