+++ de-lib/src/main/java/org/iplantc/de/collaborators/client/gin/CollaboratorsGinModule.java
@@ -3,15 +3,13 @@ package org.iplantc.de.collaborators.client.gin;
 import org.iplantc.de.client.services.GroupServiceFacade;
 import org.iplantc.de.client.services.impl.GroupServiceFacadeImpl;
 import org.iplantc.de.collaborators.client.GroupDetailsView;
-import org.iplantc.de.collaborators.client.GroupView;
 import org.iplantc.de.collaborators.client.ManageCollaboratorsView;
 import org.iplantc.de.collaborators.client.presenter.GroupDetailsPresenterImpl;
 import org.iplantc.de.collaborators.client.presenter.ManageCollaboratorsPresenter;
-import org.iplantc.de.collaborators.client.views.GroupDetailsViewImpl;
 import org.iplantc.de.collaborators.client.util.UserSearchField;
-import org.iplantc.de.collaborators.client.views.GroupViewImpl;
-import org.iplantc.de.collaborators.client.views.dialogs.ManageCollaboratorsDialog;
+import org.iplantc.de.collaborators.client.views.GroupDetailsViewImpl;
 import org.iplantc.de.collaborators.client.views.ManageCollaboratorsViewImpl;
+import org.iplantc.de.collaborators.client.views.dialogs.ManageCollaboratorsDialog;
 
 import com.google.gwt.inject.client.AbstractGinModule;
 import com.google.gwt.inject.client.assistedinject.GinFactoryModuleBuilder;
@@ -27,7 +25,6 @@ public class CollaboratorsGinModule extends AbstractGinModule {
                                              .build(ManageCollaboratorsViewFactory.class));
         bind(ManageCollaboratorsView.Presenter.class).to(ManageCollaboratorsPresenter.class);
         bind(ManageCollaboratorsDialog.class);
-        bind(GroupView.class).to(GroupViewImpl.class);
         bind(GroupServiceFacade.class).to(GroupServiceFacadeImpl.class);
         bind(GroupDetailsView.class).to(GroupDetailsViewImpl.class);
         bind(GroupDetailsView.Presenter.class).to(GroupDetailsPresenterImpl.class);
