--- de-lib/src/main/java/org/iplantc/de/collaborators/client/views/GroupViewImpl.ui.xml
@@ -1,53 +1,0 @@
-<!DOCTYPE ui:UiBinder SYSTEM "http://dl.google.com/gwt/DTD/xhtml.ent">
-<ui:UiBinder xmlns:ui="urn:ui:com.google.gwt.uibinder"
-             xmlns:container="urn:import:com.sencha.gxt.widget.core.client.container"
-             xmlns:toolbar="urn:import:com.sencha.gxt.widget.core.client.toolbar"
-             xmlns:grid="urn:import:com.sencha.gxt.widget.core.client.grid"
-             xmlns:button="urn:import:com.sencha.gxt.widget.core.client.button"
->
-
-    <ui:with field="appearance"
-             type="org.iplantc.de.collaborators.client.GroupView.GroupViewAppearance"/>
-
-    <ui:with field="margins"
-             type="com.sencha.gxt.core.client.util.Margins">
-        <ui:attributes top="0" right="0" bottom="0" left="0"/>
-    </ui:with>
-    <ui:with field="layoutData"
-             type="com.sencha.gxt.widget.core.client.container.VerticalLayoutContainer.VerticalLayoutData">
-        <ui:attributes width="1" height="-1" margins="{margins}"/>
-    </ui:with>
-
-    <ui:with type="com.sencha.gxt.widget.core.client.grid.ColumnModel"
-             field="cm"/>
-    <ui:with type="com.sencha.gxt.data.shared.ListStore"
-             field="listStore"/>
-    <ui:with type="com.sencha.gxt.widget.core.client.grid.GridView"
-             field="view">
-        <ui:attributes stripeRows="true" forceFit="true" autoFill="true"/>
-    </ui:with>
-
-
-    <container:VerticalLayoutContainer>
-        <container:child layoutData="{layoutData}">
-            <toolbar:ToolBar ui:field="toolBar">
-                <toolbar:child>
-                    <button:TextButton ui:field="addGroup"
-                                       text="{appearance.addGroup}"
-                                       icon="{appearance.addIcon}"/>
-                </toolbar:child>
-                <toolbar:child>
-                    <button:TextButton ui:field="deleteGroup"
-                                       text="{appearance.deleteGroup}"
-                                       icon="{appearance.deleteIcon}"/>
-                </toolbar:child>
-            </toolbar:ToolBar>
-        </container:child>
-        <container:child layoutData="{layoutData}">
-            <grid:Grid ui:field="grid"
-                       cm="{cm}"
-                       store="{listStore}"
-                       view="{view}"/>
-        </container:child>
-    </container:VerticalLayoutContainer>
-</ui:UiBinder>
