+++ de-lib/src/main/java/org/iplantc/de/collaborators/client/views/ManageCollaboratorsViewImpl.java
@@ -1,11 +1,8 @@
 package org.iplantc.de.collaborators.client.views;
 
 import org.iplantc.de.client.models.collaborators.Subject;
-import org.iplantc.de.client.models.groups.Group;
-import org.iplantc.de.collaborators.client.GroupView;
 import org.iplantc.de.collaborators.client.ManageCollaboratorsView;
 import org.iplantc.de.collaborators.client.events.AddGroupSelected;
-import org.iplantc.de.collaborators.client.events.DeleteGroupSelected;
 import org.iplantc.de.collaborators.client.events.GroupNameSelected;
 import org.iplantc.de.collaborators.client.events.RemoveCollaboratorSelected;
 import org.iplantc.de.collaborators.client.events.UserSearchResultSelected;
@@ -15,8 +12,6 @@ import org.iplantc.de.collaborators.shared.CollaboratorsModule;
 import org.iplantc.de.commons.client.widgets.DETabPanel;
 
 import com.google.common.base.Strings;
-import com.google.common.collect.Iterables;
-import com.google.common.collect.Lists;
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.event.shared.HandlerRegistration;
 import com.google.gwt.uibinder.client.UiBinder;
@@ -61,7 +56,6 @@ public class ManageCollaboratorsViewImpl extends Composite implements ManageColl
     @UiField BorderLayoutContainer con;
     @UiField TextButton deleteBtn;
     @UiField TextButton addGroup;
-    @UiField TextButton deleteGroup;
     @UiField Grid<Subject> grid;
     @UiField TextButton manageBtn;
     @UiField(provided = true) UserSearchField searchField;
@@ -69,7 +63,6 @@ public class ManageCollaboratorsViewImpl extends Composite implements ManageColl
     @UiField ToolBar toolbar;
     private DETabPanel tabPanel;
     @UiField FramedPanel collaboratorListPnl;
-    private GroupView groupView;
 
     @UiField(provided = true) ManageCollaboratorsView.Appearance appearance;
 
@@ -81,10 +74,8 @@ public class ManageCollaboratorsViewImpl extends Composite implements ManageColl
     @Inject
     public ManageCollaboratorsViewImpl(@Assisted final MODE mode,
                                        ManageCollaboratorsView.Appearance appearance,
-                                       GroupView groupView,
                                        UserSearchField searchField) {
         this.appearance = appearance;
-        this.groupView = groupView;
         this.searchField = searchField;
         checkBoxModel = new CheckBoxSelectionModel<>(new IdentityValueProvider<Subject>());
         initWidget(uiBinder.createAndBindUi(this));
@@ -114,50 +105,18 @@ public class ManageCollaboratorsViewImpl extends Composite implements ManageColl
     }
 
     @Override
-    public boolean hasCollaboratorsTabSelected() {
-        return tabPanel.getActiveWidget() != groupView;
-    }
-
-    @Override
     public MODE getMode() {
         return mode;
     }
 
     @Override
     public List<Subject> getSelectedSubjects() {
-        List<Group> selectedCollabLists = groupView.getSelectedCollabLists();
-        List<Subject> selectedCollaborators = grid.getSelectionModel().getSelectedItems();
-        return Lists.newArrayList(Iterables.concat(selectedCollabLists, selectedCollaborators));
+        return grid.getSelectionModel().getSelectedItems();
     }
 
-    @Override
-    public void addCollabLists(List<Group> result) {
-        groupView.addCollabLists(result);
-    }
-
-    @Override
-    public void removeCollabList(Group result) {
-        groupView.removeCollabList(result);
-    }
-
-    @Override
-    public void maskCollabLists(String loadingMask) {
-        groupView.mask(loadingMask);
-    }
-
-    @Override
-    public void unmaskCollabLists() {
-        groupView.unmask();
-    }
-    
     @Override
     public void updateCollabList(Subject group) {
         listStore.update(group);
-    }
-
-    @Override
-    public List<Group> getSelectedCollaboratorLists() {
-        return groupView.getSelectedCollabLists();
     }
 
     @Override
@@ -212,7 +171,6 @@ public class ManageCollaboratorsViewImpl extends Composite implements ManageColl
     @Override
     public void setMode(MODE mode) {
         this.mode = mode;
-        groupView.setMode(mode);
         switch (mode) {
             case MANAGE:
                 grid.getView().setEmptyText(appearance.noCollaborators());
@@ -241,11 +199,9 @@ public class ManageCollaboratorsViewImpl extends Composite implements ManageColl
         this.baseID = baseID;
         deleteBtn.ensureDebugId(baseID + CollaboratorsModule.Ids.DELETE);
         addGroup.ensureDebugId(baseID + CollaboratorsModule.Ids.ADD_GROUP);
-        deleteGroup.ensureDebugId(baseID + CollaboratorsModule.Ids.DELETE_GROUP);
         //Checkbox column config is at index 0
         grid.getView().getHeader().getHead(0).getElement().setId(baseID + CollaboratorsModule.Ids.CHECKBOX_HEADER);
         searchField.setViewDebugId(CollaboratorsModule.Ids.SEARCH_LIST);
-        groupView.asWidget().ensureDebugId(baseID + CollaboratorsModule.Ids.GROUPS_VIEW);
     }
 
     void setGridCheckBoxDebugIds() {
@@ -275,12 +231,6 @@ public class ManageCollaboratorsViewImpl extends Composite implements ManageColl
         fireEvent(new AddGroupSelected());
     }
 
-    @UiHandler("deleteGroup")
-    void deleteGroupSelected(SelectEvent event) {
-        GWT.log("Selected item " + grid.getSelectionModel().getSelectedItem().getName());
-        fireEvent(new DeleteGroupSelected(grid.getSelectionModel().getSelectedItem()));
-    }
-
     @UiHandler("manageBtn")
     void manageCollaborators(SelectEvent event) {
         setMode(MODE.MANAGE);
@@ -293,11 +243,6 @@ public class ManageCollaboratorsViewImpl extends Composite implements ManageColl
     @Override
     public HandlerRegistration addRemoveCollaboratorSelectedHandler(RemoveCollaboratorSelected.RemoveCollaboratorSelectedHandler handler) {
         return addHandler(handler, RemoveCollaboratorSelected.TYPE);
-    }
-
-    @Override
-    public HandlerRegistration addDeleteGroupSelectedHandler(DeleteGroupSelected.DeleteGroupSelectedHandler handler) {
-        return addHandler(handler, DeleteGroupSelected.TYPE);
     }
 
     @Override
