+++ de-lib/src/main/java/org/iplantc/de/collaborators/client/views/dialogs/GroupDetailsDialog.java
@@ -2,7 +2,7 @@ package org.iplantc.de.collaborators.client.views.dialogs;
 
 import org.iplantc.de.client.models.collaborators.Subject;
 import org.iplantc.de.collaborators.client.GroupDetailsView;
-import org.iplantc.de.collaborators.client.GroupView;
+import org.iplantc.de.collaborators.client.ManageCollaboratorsView;
 import org.iplantc.de.collaborators.client.events.GroupSaved;
 import org.iplantc.de.collaborators.shared.CollaboratorsModule;
 import org.iplantc.de.commons.client.views.dialogs.IPlantDialog;
@@ -21,12 +21,12 @@ import com.sencha.gxt.widget.core.client.event.SelectEvent;
 public class GroupDetailsDialog extends IPlantDialog implements GroupSaved.HasGroupSavedHandlers {
 
     GroupDetailsView.Presenter presenter;
-    GroupView.GroupViewAppearance appearance;
+    ManageCollaboratorsView.Appearance appearance;
     GroupDetailsView.MODE mode = GroupDetailsView.MODE.EDIT;
 
     @Inject
     public GroupDetailsDialog(GroupDetailsView.Presenter presenter,
-                              GroupView.GroupViewAppearance appearance) {
+                              ManageCollaboratorsView.Appearance appearance) {
         this.presenter = presenter;
         this.appearance = appearance;
