+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/collaborators/Collaborators.gwt.xml
@@ -8,14 +8,6 @@
         <when-type-is class="org.iplantc.de.collaborators.client.ManageCollaboratorsView.Appearance" />
     </replace-with>
 
-    <replace-with class="org.iplantc.de.theme.base.client.collaborators.GroupViewDefaultAppearance">
-        <when-type-is class="org.iplantc.de.collaborators.client.GroupView.GroupViewAppearance"/>
-    </replace-with>
-
-    <replace-with class="org.iplantc.de.theme.base.client.collaborators.cells.GroupNameCellDefaultAppearance">
-        <when-type-is class="org.iplantc.de.collaborators.client.views.cells.GroupNameCell.GroupNameCellAppearance"/>
-    </replace-with>
-
     <replace-with class="org.iplantc.de.theme.base.client.collaborators.cells.SubjectNameCellDefaultAppearance">
         <when-type-is class="org.iplantc.de.collaborators.client.views.cells.SubjectNameCell.SubjectNameCellAppearance"/>
     </replace-with>
