+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/collaborators/CollaboratorsDisplayStrings.java
@@ -2,6 +2,8 @@ package org.iplantc.de.theme.base.client.collaborators;
 
 import com.google.gwt.i18n.client.Messages;
 
+import java.util.List;
+
 /**
  * @author aramsey
  */
@@ -14,4 +16,30 @@ public interface CollaboratorsDisplayStrings extends Messages {
     String collaboratorTab();
 
     String collaboratorListTab();
+
+    String groupNameLabel();
+
+    String groupDescriptionLabel();
+
+    String newGroupDetailsHeading();
+
+    String editGroupDetailsHeading(String name);
+
+    String deleteGroupConfirmHeading(String name);
+
+    String deleteGroupConfirm(String name);
+
+    String unableToAddMembers(@PluralCount List<String> memberString);
+
+    String groupCreatedSuccess(String name);
+
+    String memberDeleteFail(@PluralCount List<String> memberNames);
+
+    String groupSelfAdd();
+
+    String memberAddToGroupsSuccess(String subjectDisplayName);
+
+    String groupNameValidationMsg(String restrictedChars);
+
+    String addGroup();
 }
