--- de-lib/src/main/java/org/iplantc/de/theme/base/client/collaborators/GroupDisplayStrings.properties
@@ -1,18 +1,0 @@
-addGroup=Add List
-deleteGroup=Delete List
-deleteGroupConfirm=By deleting the collaborator list {0}, any data that has been shared with this list will immediately be un-shared.  Do you wish to continue deleting the collaborator list?
-deleteGroupConfirmHeading=Delete {0} Collaborator List?
-editGroupDetailsHeading=Edit {0}
-groupCreatedSuccess=Successfully created {0} Collaborator List
-groupDeleteSuccess=Successfully deleted {0} Collaborator List
-groupDescriptionLabel=Description
-groupNameLabel=Collaborator List Name
-groupNameValidationMsg=The following characters are not supported: {0}
-groupSelfAdd=Cannot add list as a member to itself
-memberAddSuccess={0} successfully added to {1}
-memberAddToGroupsSuccess={0} successfully added to selected lists
-memberDeleteFail=Unable to remove the following collaborators: {0,list}
-newGroupDetailsHeading=Create a Collaborator List
-noCollabLists = Click the Add button to start creating Collaborator Lists
-noCollabListSelected=No Collaborator List selected to add collaborator to
-unableToAddMembers=Unable to add the following members: {0,list}
