+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/collaborators/ManageCollaboratorsViewDefaultAppearance.java
@@ -1,10 +1,16 @@
 package org.iplantc.de.theme.base.client.collaborators;
 
+import org.iplantc.de.client.models.collaborators.Subject;
+import org.iplantc.de.client.models.groups.Group;
+import org.iplantc.de.client.models.groups.UpdateMemberResult;
 import org.iplantc.de.collaborators.client.ManageCollaboratorsView;
 import org.iplantc.de.resources.client.IplantResources;
 import org.iplantc.de.resources.client.messages.IplantContextualHelpStrings;
 import org.iplantc.de.resources.client.messages.IplantDisplayStrings;
+import org.iplantc.de.resources.client.messages.IplantErrorStrings;
+import org.iplantc.de.resources.client.messages.IplantValidationMessages;
 
+import com.google.common.base.Strings;
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.resources.client.ImageResource;
 import com.google.gwt.safehtml.shared.SafeHtml;
@@ -13,6 +19,9 @@ import com.sencha.gxt.core.client.XTemplates;
 import com.sencha.gxt.theme.base.client.grid.CheckBoxColumnDefaultAppearance.CheckBoxColumnResources;
 import com.sencha.gxt.theme.base.client.grid.CheckBoxColumnDefaultAppearance.CheckBoxColumnStyle;
 import com.sencha.gxt.theme.base.client.grid.CheckBoxColumnDefaultAppearance.CheckBoxColumnTemplates;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * @author jstroot
@@ -26,6 +35,9 @@ public class ManageCollaboratorsViewDefaultAppearance implements ManageCollabora
     public interface CheckBoxColumnDebugTemplates extends XTemplates {
        @XTemplates.XTemplate("<div id='{debugId}' class='{style.hdChecker}'></div>")
         SafeHtml renderDebugHeader(CheckBoxColumnStyle style, String debugId);
+
+        @XTemplates.XTemplate("<span style='color: red;'>*&nbsp</span>{label}")
+        SafeHtml requiredFieldLabel(String label);
     }
 
     private final CheckBoxColumnResources resources;
@@ -35,6 +47,8 @@ public class ManageCollaboratorsViewDefaultAppearance implements ManageCollabora
     private IplantContextualHelpStrings iplantContextualHelpStrings;
     private IplantResources iplantResources;
     private CollaboratorsDisplayStrings displayStrings;
+    private IplantErrorStrings iplantErrorStrings;
+    private IplantValidationMessages iplantValidationMessages;
 
     public ManageCollaboratorsViewDefaultAppearance() {
         this(GWT.<CheckBoxColumnResources> create(CheckBoxColumnResources.class),
@@ -42,7 +56,9 @@ public class ManageCollaboratorsViewDefaultAppearance implements ManageCollabora
              GWT.<IplantDisplayStrings>create(IplantDisplayStrings.class),
              GWT.<IplantContextualHelpStrings>create(IplantContextualHelpStrings.class),
              GWT.<IplantResources> create(IplantResources.class),
-             GWT.<CollaboratorsDisplayStrings> create(CollaboratorsDisplayStrings.class));
+             GWT.<CollaboratorsDisplayStrings> create(CollaboratorsDisplayStrings.class),
+             GWT.<IplantErrorStrings> create(IplantErrorStrings.class),
+             GWT.<IplantValidationMessages> create(IplantValidationMessages.class));
     }
 
     ManageCollaboratorsViewDefaultAppearance(final CheckBoxColumnResources resources,
@@ -50,7 +66,9 @@ public class ManageCollaboratorsViewDefaultAppearance implements ManageCollabora
                                              IplantDisplayStrings iplantDisplayStrings,
                                              IplantContextualHelpStrings iplantContextualHelpStrings,
                                              IplantResources iplantResources,
-                                             CollaboratorsDisplayStrings displayStrings) {
+                                             CollaboratorsDisplayStrings displayStrings,
+                                             IplantErrorStrings iplantErrorStrings,
+                                             IplantValidationMessages iplantValidationMessages) {
         this.resources = resources;
         this.style = resources.style();
         this.templates = templates;
@@ -58,6 +76,8 @@ public class ManageCollaboratorsViewDefaultAppearance implements ManageCollabora
         this.iplantContextualHelpStrings = iplantContextualHelpStrings;
         this.iplantResources = iplantResources;
         this.displayStrings = displayStrings;
+        this.iplantErrorStrings = iplantErrorStrings;
+        this.iplantValidationMessages = iplantValidationMessages;
 
         style.ensureInjected();
     }
@@ -136,5 +156,117 @@ public class ManageCollaboratorsViewDefaultAppearance implements ManageCollabora
     @Override
     public String loadingMask() {
         return iplantDisplayStrings.loadingMask();
+    }
+
+    @Override
+    public String addGroup() {
+        return displayStrings.addGroup();
+    }
+
+    @Override
+    public ImageResource addIcon() {
+        return iplantResources.add();
+    }
+
+    @Override
+    public SafeHtml groupNameLabel() {
+        return templates.requiredFieldLabel(displayStrings.groupNameLabel());
+    }
+    @Override
+    public String groupDescriptionLabel() {
+        return displayStrings.groupDescriptionLabel();
+    }
+
+    @Override
+    public int groupDetailsWidth() {
+        return 500;
+    }
+
+    @Override
+    public int groupDetailsHeight() {
+        return 500;
+    }
+
+    @Override
+    public String groupDetailsHeading(Subject subject) {
+        if (subject == null || Strings.isNullOrEmpty(subject.getName())) {
+            return displayStrings.newGroupDetailsHeading();
+        } else {
+            return displayStrings.editGroupDetailsHeading(subject.getName());
+        }
+    }
+
+    @Override
+    public String completeRequiredFieldsError() {
+        return iplantDisplayStrings.completeRequiredFieldsError();
+    }
+
+    @Override
+    public String deleteGroupConfirmHeading(Group group) {
+        return displayStrings.deleteGroupConfirmHeading(group.getName());
+    }
+
+    @Override
+    public String deleteGroupConfirm(Group group) {
+        return displayStrings.deleteGroupConfirm(group.getName());
+    }
+
+    @Override
+    public String unableToAddMembers(List<UpdateMemberResult> failures) {
+        List<String> memberNames = failures.stream().map(UpdateMemberResult::getSubjectName).collect(
+                Collectors.toList());
+        return displayStrings.unableToAddMembers(memberNames);
+    }
+
+    @Override
+    public String groupCreatedSuccess(Group group) {
+        return displayStrings.groupCreatedSuccess(group.getName());
+    }
+
+    @Override
+    public String memberDeleteFail(List<UpdateMemberResult> subjects) {
+        List<String> memberNames = subjects.stream().map(UpdateMemberResult::getSubjectName).collect(
+                Collectors.toList());
+        return displayStrings.memberDeleteFail(memberNames);
+    }
+
+    @Override
+    public String collaboratorsSelfAdd() {
+        return iplantDisplayStrings.collaboratorSelfAdd();
+    }
+
+    @Override
+    public String groupSelfAdd() {
+        return displayStrings.groupSelfAdd();
+    }
+
+    @Override
+    public String collaboratorRemoveConfirm(String names) {
+        return iplantDisplayStrings.collaboratorRemoveConfirm(names);
+    }
+
+    @Override
+    public String collaboratorAddConfirm(String names) {
+        return iplantDisplayStrings.collaboratorAddConfirm(names);
+    }
+
+    @Override
+    public String addCollabErrorMsg() {
+        return iplantErrorStrings.addCollabErrorMsg();
+    }
+
+    @Override
+    public String memberAddToGroupsSuccess(Subject subject) {
+        return displayStrings.memberAddToGroupsSuccess(subject.getSubjectDisplayName());
+    }
+
+    @Override
+    public String groupNameValidationMsg(String restrictedChars) {
+        return displayStrings.groupNameValidationMsg(restrictedChars);
+    }
+
+    @Override
+    public String invalidChars(String restrictedChars) {
+        return iplantValidationMessages.invalidChars(restrictedChars);
     }
 }
